training/requirements.txt
h5py==3.1.0
Pillow==8.0.1
pandas==1.1.4
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0+cu101
torchvision==0.8.1+cu101
torchaudio==0.7.0